requirements.txt
colorama>=0.4.0
pyyaml>=6.0
click>=8.0
